+++ packages/registry/src/index.ts
@@ -19,8 +19,10 @@ import { Optimism } from '@rabbitholegg/questdk-plugin-optimism'
 import { Hop } from '@rabbitholegg/questdk-plugin-hop'
 import { Arbitrum } from '@rabbitholegg/questdk-plugin-arbitrum'
 import { GMX } from '@rabbitholegg/questdk-plugin-gmx'
+import { Camelot } from '@rabbitholegg/questdk-plugin-camelot'
 import { Tally } from '@rabbitholegg/questdk-plugin-tally'
 import { BasePaint } from '@rabbitholegg/questdk-plugin-basepaint'
+import { ENTRYPOINT } from './contract-addresses'
 
 export const plugins: Record<string, IActionPlugin> = {
   [Connext.pluginId]: Connext,
@@ -33,6 +35,7 @@ export const plugins: Record<string, IActionPlugin> = {
   [Optimism.pluginId]: Optimism,
   [GMX.pluginId]: GMX,
   [Tally.pluginId]: Tally,
+  [Camelot.pluginId]: Camelot,
   [BasePaint.pluginId]: BasePaint,
 }
 
@@ -64,4 +67,8 @@ export const executePlugin = (
     default:
       throw new Error(`Unknown action type "${actionType}"`)
   }
+}
+
+export const getIndexedContracts = (_chainId: number) => {
+  return [ENTRYPOINT]
 }